Implement API as an alternative to Websocket
Messages from server to client can be send only via Websocket, but it is
possible to use API for client-to-server messages. It is more
convenient, but creates latency between clients.

Tests should be done to check, how latency affects user experience. For
that reason, server still processes messages from client. Client snippet
introduces variable to control, whether to use API or Websocket for
client-to-server messages.